+++ runtime/runtime_memory_metering_test.go
@@ -29,7 +29,9 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/onflow/cadence"
+	jsoncdc "github.com/onflow/cadence/encoding/json"
 	"github.com/onflow/cadence/runtime/common"
+	"github.com/onflow/cadence/runtime/parser2"
 	"github.com/onflow/cadence/runtime/tests/utils"
 )
 
@@ -966,5 +968,108 @@ func TestStorageCommitsMetering(t *testing.T) {
 		require.NoError(t, err)
 		assert.True(t, storageUsedInvoked)
 		assert.Equal(t, uint64(4), meter.getMemory(common.MemoryKindAtreeEncodedSlab))
+	})
+}
+
+func TestMemoryMeteringErrors(t *testing.T) {
+
+	t.Parallel()
+
+	runtime := newTestInterpreterRuntime()
+
+	type memoryMeter map[common.MemoryKind]uint64
+
+	runtimeInterface := func(meter memoryMeter) *testRuntimeInterface {
+		intf := &testRuntimeInterface{
+			meterMemory: func(usage common.MemoryUsage) error {
+				if usage.Kind == common.MemoryKindStringValue ||
+					usage.Kind == common.MemoryKindArrayValueBase {
+					return testMemoryError{}
+				}
+				return nil
+			},
+		}
+		intf.decodeArgument = func(b []byte, t cadence.Type) (cadence.Value, error) {
+			return jsoncdc.Decode(intf, b)
+		}
+		return intf
+	}
+
+	executeScript := func(script []byte, meter memoryMeter, args ...cadence.Value) error {
+		_, err := runtime.ExecuteScript(
+			Script{
+				Source:    script,
+				Arguments: encodeArgs(args),
+			},
+			Context{
+				Interface: runtimeInterface(meter),
+				Location:  utils.TestLocation,
+			},
+		)
+
+		return err
+	}
+
+	t.Run("no errors", func(t *testing.T) {
+		t.Parallel()
+
+		script := []byte(`
+            pub fun main() {}
+        `)
+
+		err := executeScript(script, memoryMeter{})
+		assert.NoError(t, err)
+	})
+
+	t.Run("importing", func(t *testing.T) {
+		t.Parallel()
+
+		script := []byte(`
+            pub fun main(x: String) {}
+        `)
+
+		err := executeScript(
+			script,
+			memoryMeter{},
+			cadence.String("hello"),
+		)
+
+		assert.ErrorIs(t, err, testMemoryError{})
+	})
+
+	t.Run("at lexer", func(t *testing.T) {
+		t.Parallel()
+
+		script := []byte(`
+            pub fun main() {
+                let x = "hello"
+            }
+        `)
+
+		err := executeScript(script, memoryMeter{})
+
+		require.IsType(t, Error{}, err)
+		runtimeError := err.(Error)
+
+		require.IsType(t, &ParsingCheckingError{}, runtimeError.Err)
+		parsingCheckingError := runtimeError.Err.(*ParsingCheckingError)
+
+		require.IsType(t, parser2.Error{}, parsingCheckingError.Err)
+		parserError := parsingCheckingError.Err.(parser2.Error)
+
+		assert.Contains(t, parserError.Error(), "memory limit exceeded")
+	})
+
+	t.Run("at interpreter", func(t *testing.T) {
+		t.Parallel()
+
+		script := []byte(`
+            pub fun main() {
+                let x: [AnyStruct] = []
+            }
+        `)
+
+		err := executeScript(script, memoryMeter{})
+		assert.ErrorIs(t, err, testMemoryError{})
 	})
 }
